Add Maven Shade and GraalVM plugins, update resources
The Maven Shade plugin was added to create executable JARs. Properties for the main class and new plugin versions were added to the pom file. In addition, the GraalVM Native Image plugin was included to create standalone executables. Resources were also updated to include META-INF and webapp directories, as well as a new reflect-config.json file.